requirements.txt
prettytable==0.7.2
MySQL-python==1.2.5
mysql-connector-python==1.2.3
requests==2.4.1
prettytable==0.7.2